Always do transformations on router
